[Refactor] CoreLocation 관련 공통메서드 Protocol로 분리
